@@ -1,5 +1,5 @@
 /*
- * Copyright 2024, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,8 +24,6 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import org.gradle.configurationcache.extensions.capitalized
-
 /**
  * This file provides extensions to `String` and `CharSequence` that wrap
  * analogues from standard Kotlin runtime.
@@ -43,19 +41,10 @@ private const val ABOUT = ""
 /**
  * Makes the first character come in the title case.
  */
-fun String.titleCaseFirstChar(): String {
-    // return replaceFirstChar { it.titlecase() }
-    // OR for earlier Kotlin versions:
-    //   1. add import of `org.gradle.configurationcache.extensions.capitalized`
-    //   2. call `capitalized()` instead of `replaceFirstChar` above.
-    return capitalized()
-}
+fun String.titleCaseFirstChar(): String = replaceFirstChar { it.titlecase() }
 
 /**
  * Converts this string to lowercase.
  */
-fun String.lowercased(): String {
-    //    return lowercase()
-    // OR for earlier Kotlin versions call:
-    return toLowerCase()
-}
+@Deprecated(message = "Please use `lowercase()` instead.", replaceWith = ReplaceWith("lowercase"))
+fun String.lowercased(): String = lowercase()
