@@ -42,12 +42,12 @@ object McJava {
     /**
      * The version used to in the build classpath.
      */
-    const val dogfoodingVersion = "2.0.0-SNAPSHOT.264"
+    const val dogfoodingVersion = "2.0.0-SNAPSHOT.266"
 
     /**
      * The version to be used for integration tests.
      */
-    const val version = "2.0.0-SNAPSHOT.265"
+    const val version = "2.0.0-SNAPSHOT.266"
 
     /**
      * The ID of the Gradle plugin.
