@@ -1,5 +1,5 @@
 /*
- * Copyright 2024, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -49,8 +49,6 @@ fun JsPlugins.protobuf() {
 
     val protobufExt = project.extensions.getByType(ProtobufExtension::class.java)
     protobufExt.apply {
-
-        generatedFilesBaseDir = projectDir.path
 
         protoc {
             artifact = Protobuf.compiler
